+++ NEWS.md
@@ -16,6 +16,7 @@
     - `$str$strip()` -> `$str$strip_chars()`
     - `$str$lstrip()` -> `$str$strip_chars_start()`
     - `$str$rstrip()` -> `$str$strip_chars_end()`
+- Argument `pattern` of `$str$count_matches()` can now use expressions.
 - `$get_optimization_toggle()` now returns two additional values: `fast_projection`
   and `eager`.
 - Updated Rust toolchain.